Handle CredentialRetrievalError errors from boto3
Seen this once, and I think (hope :)) it's some temporary issue.  It did
cause the SQS poller service to die, but after restarting buildbot it
came back to life all happy and cheerful.

If so, fixing like this could be fine for now. For robustness, one could
consider putting the service in some failure mode with exponential
backoff. I don't think it will become a big issue for now with one
minute polling interval.