License compliance!
We have taken some software from buildbot (to be able to run twisted
tests), and we we're not compliant. (Copyright statements intact, but
not the license). Turns out, we took stuff with two licenses: GPLv2 and
MIT. We can be comliant by putting the GPL code in a GPL module and the
MIT in a MIT module. Because our unit tests depend on (and thus derives
from) a GPL licenses work, the unit tests themselves are also GPLv2.

The reason for vendoring this was that (at some point, haven't verified
this is still the case) buildbot's testing infrastructure wasn't
included in the release tarballs. If we can simply import from the
buildbot codebase, we can remove the vendored test support.

All other code is (for now) authored by me, and the license is currently
undetermined (but likely will be MIT). Let me now if you need/want me to
prioritize this formality.